requirements.txt
streamlit
torch
torchvision
numpy
pandas
matplotlib
pillow
